Create Split, read, write
Assign name to the file, assign remove lost to a list of ip address that are no longer allowed to access restricted information. Use split to convert ip addresses from a string to a list. Then build iterative statement, name loop variable 'element', loop through 'ip addresses. Build conditional statement, if current element is in 'remove list'. Convert 'ip addresses back to a string so that it can be written into the text file.  Then build 'with' statement to rewrite the original file. Rewrite the file, replacing its contents with 'ip addresses. Build 'with' statement to read in the updated file, read in the updated file and store the contents in 'text. Then print the text.